+++ parentroadmap/templates/base.html
@@ -5,7 +5,7 @@
 <head>
     <meta charset="UTF-8">
     <title>{{title}}</title>
-    <link type="text/css" href="{% static 'parentroadmapapp/css/styles.css' %}" rel="stylesheet" />
+    <link type="text/css" href="{% static 'webApp/css/styles.css' %}" rel="stylesheet" />
 </head>
 <body>
 <tr><td valign=top>
