+++ parentroadmap/webApp/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ParentroadmapappConfig(AppConfig):
+class WebAppConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'parentroadmapapp'
+    name = 'webApp'
