+++ parentroadmap/webProject/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for parentroadmap project.
+Django settings for webProject webProject.
 
 Generated by 'django-admin startproject' using Django 4.2.1.
 
@@ -12,7 +12,7 @@ https://docs.djangoproject.com/en/4.2/ref/settings/
 
 from pathlib import Path
 
-# Build paths inside the project like this: BASE_DIR / 'subdir'.
+# Build paths inside the webProject like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
 
 
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'parentroadmapapp.apps.ParentroadmapappConfig',
+    'webApp.apps.WebAppConfig',
 ]
 
 MIDDLEWARE = [
@@ -50,7 +50,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'parentroadmap.urls'
+ROOT_URLCONF = 'webProject.urls'
 
 TEMPLATES = [
     {
@@ -70,7 +70,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'parentroadmap.wsgi.application'
+WSGI_APPLICATION = 'webProject.wsgi.application'
 
 
 # Database
