+++ parentroadmap/webProject/urls.py
@@ -1,5 +1,5 @@
 """
-URL configuration for parentroadmap project.
+URL configuration for webProject webProject.
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/4.2/topics/http/urls/
@@ -19,5 +19,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('parentroadmapapp.urls')),
+    path('', include('webApp.urls')),
 ]
